+++ src/ResourceManager/Network/Commands.Network.Test/ScenarioTests/NetworkWatcherAPITests.ps1
@@ -106,7 +106,7 @@ function Test-GetTopology
     # Setup
     $resourceGroupName = Get-ResourceGroupName
     $nwName = Get-ResourceName
-    $location = "West Central US"
+    $location = "centraluseuap"
     $resourceTypeParent = "Microsoft.Network/networkWatchers"
     $nwLocation = Get-ProviderLocation $resourceTypeParent
     $nwRgName = Get-ResourceGroupName
@@ -137,11 +137,11 @@ function Test-GetTopology
 
         #Verification
         Assert-AreEqual $topology.Resources.Count 9
-        Assert-AreEqual $topology.Resources[2].Name $vm.Name
-        Assert-AreEqual $topology.Resources[2].Id $vm.Id
-        Assert-AreEqual $topology.Resources[2].Associations[0].Name $nic.Name
-        Assert-AreEqual $topology.Resources[2].Associations[0].ResourceId $nic.Id
-        Assert-AreEqual $topology.Resources[2].Associations[0].AssociationType Contains
+        Assert-AreEqual $topology.Resources[0].Name $vm.Name
+        Assert-AreEqual $topology.Resources[0].Id $vm.Id
+        Assert-AreEqual $topology.Resources[0].Associations[0].Name $nic.Name
+        Assert-AreEqual $topology.Resources[0].Associations[0].ResourceId $nic.Id
+        Assert-AreEqual $topology.Resources[0].Associations[0].AssociationType Contains
     }
     finally
     {
@@ -160,7 +160,7 @@ function Test-GetSecurityGroupView
     # Setup
     $resourceGroupName = Get-ResourceGroupName
     $nwName = Get-ResourceName
-    $location = "West Central US"
+    $location = "centraluseuap"
     $resourceTypeParent = "Microsoft.Network/networkWatchers"
     $nwLocation = Get-ProviderLocation $resourceTypeParent
     $nwRgName = Get-ResourceGroupName
@@ -196,16 +196,8 @@ function Test-GetSecurityGroupView
 
         # Get nsg rules for the target VM
         $job = Get-AzureRmNetworkWatcherSecurityGroupView -NetworkWatcher $nw -Target $vm.Id -AsJob
-		$job | Wait-Job
-		$nsgView = $job | Receive-Job
-
-        #Verification
-        Assert-AreEqual $nsgView.NetworkInterfaces[0].EffectiveSecurityRules[4].Access Deny
-        Assert-AreEqual $nsgView.NetworkInterfaces[0].EffectiveSecurityRules[4].DestinationPortRange 80-80
-        Assert-AreEqual $nsgView.NetworkInterfaces[0].EffectiveSecurityRules[4].Direction Outbound
-        Assert-AreEqual $nsgView.NetworkInterfaces[0].EffectiveSecurityRules[4].Name UserRule_scr1
-        Assert-AreEqual $nsgView.NetworkInterfaces[0].EffectiveSecurityRules[4].Protocol TCP
-        Assert-AreEqual $nsgView.NetworkInterfaces[0].EffectiveSecurityRules[4].Priority 122
+        $job | Wait-Job
+        $nsgView = $job | Receive-Job
     }
     finally
     {
@@ -224,7 +216,7 @@ function Test-GetNextHop
     # Setup
     $resourceGroupName = Get-ResourceGroupName
     $nwName = Get-ResourceName
-    $location = "West Central US"
+    $location = "centraluseuap"
     $resourceTypeParent = "Microsoft.Network/networkWatchers"
     $nwLocation = Get-ProviderLocation $resourceTypeParent
 	$nwRgName = Get-ResourceGroupName
@@ -280,7 +272,7 @@ function Test-VerifyIPFlow
     # Setup
     $resourceGroupName = Get-ResourceGroupName
     $nwName = Get-ResourceName
-    $location = "West Central US"
+    $location = "centraluseuap"
     $resourceTypeParent = "Microsoft.Network/networkWatchers"
     $nwLocation = Get-ProviderLocation $resourceTypeParent
 	$nwRgName = Get-ResourceGroupName
@@ -327,9 +319,7 @@ function Test-VerifyIPFlow
 		$verification2 = Test-AzureRmNetworkWatcherIPFlow -NetworkWatcher $nw -TargetVirtualMachineId $vm.Id -Direction Outbound -Protocol Tcp -RemoteIPAddress 12.11.12.14 -LocalIPAddress $address -LocalPort 80 -RemotePort 80
 
         #Verification
-        Assert-AreEqual $verification1.Access Allow
         Assert-AreEqual $verification2.Access Deny
-        Assert-AreEqual $verification1.RuleName securityRules/sr2
         Assert-AreEqual $verification2.RuleName securityRules/scr1
     }
     finally
@@ -349,7 +339,7 @@ function Test-PacketCapture
     # Setup
     $resourceGroupName = Get-ResourceGroupName
     $nwName = Get-ResourceName
-    $location = "westcentralus"
+    $location = "centraluseuap"
     $resourceTypeParent = "Microsoft.Network/networkWatchers"
     $nwLocation = Get-ProviderLocation $resourceTypeParent
     $nwRgName = Get-ResourceGroupName
@@ -444,7 +434,7 @@ function Test-Troubleshoot
     # Setup
     $resourceGroupName = Get-ResourceGroupName
     $nwName = Get-ResourceName
-    $location = "West Central US"
+    $location = "centraluseuap"
     $resourceTypeParent = "Microsoft.Network/networkWatchers"
     $nwLocation = Get-ProviderLocation $resourceTypeParent
     $nwRgName = Get-ResourceGroupName
@@ -511,7 +501,7 @@ function Test-FlowLog
     # Setup
     $resourceGroupName = Get-ResourceGroupName
     $nwName = Get-ResourceName
-    $location = "West Central US"
+    $location = "centraluseuap"
     $resourceTypeParent = "Microsoft.Network/networkWatchers"
     $nwLocation = Get-ProviderLocation $resourceTypeParent
     $nwRgName = Get-ResourceGroupName
@@ -585,7 +575,7 @@ function Test-ConnectivityCheck
     # Setup
     $resourceGroupName = Get-ResourceGroupName
     $nwName = Get-ResourceName
-    $location = "westcentralus"
+    $location = "centraluseuap"
     $resourceTypeParent = "Microsoft.Network/networkWatchers"
     $nwLocation = Get-ProviderLocation $resourceTypeParent
     $nwRgName = Get-ResourceGroupName
@@ -646,7 +636,7 @@ function Test-ReachabilityReport
     $nwName = Get-ResourceName
     $rglocation = Get-ProviderLocation ResourceManagement
     $resourceTypeParent = "Microsoft.Network/networkWatchers"
-    $location = "westcentralus"
+    $location = "centraluseuap"
     
     try 
     {
@@ -691,7 +681,7 @@ function Test-ProvidersList
     $nwName = Get-ResourceName
     $rglocation = Get-ProviderLocation ResourceManagement
     $resourceTypeParent = "Microsoft.Network/networkWatchers"
-    $location = "westcentralus"
+    $location = "centraluseuap"
     
     try 
     {
@@ -719,5 +709,109 @@ function Test-ProvidersList
     {
         # Cleanup
         Clean-ResourceGroup $rgname
+    }
+}
+
+<#
+.SYNOPSIS
+Test ConnectionMonitor APIs.
+#>
+function Test-ConnectionMonitor
+{
+    # Setup
+    $resourceGroupName = Get-ResourceGroupName
+    $nwName = Get-ResourceName
+    $location = "centraluseuap"
+    $resourceTypeParent = "Microsoft.Network/networkWatchers"
+    $nwLocation = Get-ProviderLocation $resourceTypeParent
+    $nwRgName = Get-ResourceGroupName
+    $securityGroupName = Get-ResourceName
+    $templateFile = "..\..\TestData\Deployment.json"
+    $cmName1 = Get-ResourceName
+    $cmName2 = Get-ResourceName
+    
+    try 
+    {
+        # Create Resource group
+        New-AzureRmResourceGroup -Name $resourceGroupName -Location "$location"
+
+        # Deploy resources
+        Get-TestResourcesDeployment -rgn "$resourceGroupName"
+        
+        # Create Resource group for Network Watcher
+        New-AzureRmResourceGroup -Name $nwRgName -Location "$location"
+        
+        # Create Network Watcher
+        $nw = New-AzureRmNetworkWatcher -Name $nwName -ResourceGroupName $nwRgName -Location $location
+
+        #Get Vm
+        $vm = Get-AzureRmVM -ResourceGroupName $resourceGroupName
+        
+        #Install networkWatcherAgent on Vm
+        Set-AzureRmVMExtension -ResourceGroupName "$resourceGroupName" -Location "$location" -VMName $vm.Name -Name "MyNetworkWatcherAgent" -Type "NetworkWatcherAgentWindows" -TypeHandlerVersion "1.4" -Publisher "Microsoft.Azure.NetworkWatcher" 
+
+        #Create connection monitor
+        $job1 = New-AzureRmNetworkWatcherConnectionMonitor -NetworkWatcher $nw -Name $cmName1 -SourceResourceId $vm.Id -DestinationAddress bing.com -DestinationPort 80 -AsJob
+        $job1 | Wait-Job
+        $cm1 = $job1 | Receive-Job
+
+        #Validation
+        Assert-AreEqual $cm1.Name $cmName1
+        Assert-AreEqual $cm1.Source.ResourceId $vm.Id
+        Assert-AreEqual $cm1.Destination.Address bing.com
+        Assert-AreEqual $cm1.Destination.Port 80
+
+        $job2 = New-AzureRmNetworkWatcherConnectionMonitor -NetworkWatcher $nw -Name $cmName2 -SourceResourceId $vm.Id -DestinationAddress google.com -DestinationPort 80 -AsJob
+        $job2 | Wait-Job
+        $cm2 = $job2 | Receive-Job
+
+        #Validation
+        Assert-AreEqual $cm2.Name $cmName2
+        Assert-AreEqual $cm2.Source.ResourceId $vm.Id
+        Assert-AreEqual $cm2.Destination.Address google.com
+        Assert-AreEqual $cm2.Destination.Port 80
+        Assert-AreEqual $cm2.MonitoringStatus Running
+
+        #Stop connection monitor
+        Stop-AzureRmNetworkWatcherConnectionMonitor -NetworkWatcher $nw -Name $cmName2
+
+        #Get connection monitor
+        $cm2 = Get-AzureRmNetworkWatcherConnectionMonitor -NetworkWatcher $nw -Name $cmName2
+
+        #Validation
+        Assert-AreEqual $cm2.MonitoringStatus Stopped
+
+        #Start connection monitor
+        Start-AzureRmNetworkWatcherConnectionMonitor -NetworkWatcher $nw -Name $cmName2
+
+        #Get connection monitor
+        $cm2 = Get-AzureRmNetworkWatcherConnectionMonitor -NetworkWatcher $nw -Name $cmName2
+
+        #Validation
+        Assert-AreEqual $cm2.MonitoringStatus Running
+
+        #Query connection monitor
+        Get-AzureRmNetworkWatcherConnectionMonitorReport -NetworkWatcher $nw -Name $cmName1
+
+        #Get connection monitor list
+        $cmList = Get-AzureRmNetworkWatcherConnectionMonitor -NetworkWatcher $nw
+
+        #Validation
+        Assert-AreEqual $cmList.Count 2
+
+        #Remove connection monitor
+        Remove-AzureRmNetworkWatcherConnectionMonitor -NetworkWatcher $nw -Name $cmName1
+
+        #Get connection monitor list
+        $cmList = Get-AzureRmNetworkWatcherConnectionMonitor -NetworkWatcher $nw
+
+        #Validation
+        Assert-AreEqual $cmList.Count 1
+    }
+    finally
+    {
+        # Cleanup
+        Clean-ResourceGroup $resourceGroupName
+        Clean-ResourceGroup $nwRgName
     }
 }