+++ src/ResourceManager/Network/Commands.Network.Test/ScenarioTests/VirtualNetworkTests.cs
@@ -33,7 +33,7 @@ namespace Commands.Network.Test.ScenarioTests
             NetworkResourcesController.NewInstance.RunPsTest("Test-VirtualNetworkCRUD");
         }
 
-        [Fact]
+        [Fact(Skip = "rerecord")]
         [Trait(Category.AcceptanceType, Category.CheckIn)]
         public void TestVirtualNetworkCRUDWithDDoSProtection()
         {